clinic debug for test added

+++ life-index/src/main/java/spb/hack/lifeindex/model/dto/RequestParamsDto.java
@@ -14,7 +14,7 @@ public class RequestParamsDto {
     // всевозможные данные, которые мы передаём в запросы
     private House house;
     private String address;
-    private Long radius;
+    private Integer radius;
     private String modifier;
     private Integer page;
 }
